refactor: propogate range values
Previously we were searching every end location and checking if it
appeared in the starting seed set after mapping. As each contiguous
range that is affected by the same transforms will end up as a
contiguous block, we only need to check the minimum value in each
block. Changed the calculations to use the initial seed ranges and
apply all changes through the maps, splitting them as they fall in
different combinations of transforms. Then we only need to check the
minimum value of each range. Resulted in a ~1000x speedup.